requirements.txt
numpy
scipy
scikit-learn == 0.19.1
torch == 0.4.0
python-weka-wrapper3
